==> 12.R <==
library(limma)
lncFile <- "diffLncExp.txt"
cliFile <- "time.txt"
setwd("C:\\Users\\Zhang TONGTONG\\Desktop\\3\\12-1")

rt <- read.table(lncFile, header = T, sep = "\t", check.names = F)
rt <- as.matrix(rt)
rownames(rt) <- rt[, 1]
exp <- rt[, 2:ncol(rt)]
dimnames <- list(rownames(exp), colnames(exp))
data <- matrix(as.numeric(as.matrix(exp)), nrow = nrow(exp), dimnames = dimnames)
data <- avereps(data)

group <- sapply(strsplit(colnames(data), "\\-"), "[", 4)
group <- sapply(strsplit(group, ""), "[", 1)
group <- gsub("2", "1", group)
data <- data[, group == 0]
colnames(data) <- gsub("(.*?)\\-(.*?)\\-(.*?)\\-(.*?)\\-.*", "\\1\\-\\2\\-\\3", colnames(data))
data <- t(data)
data <- avereps(data)

cli <- read.table(cliFile, header = T, sep = "\t", check.names = F, row.names = 1)

sameSample <- intersect(row.names(data), row.names(cli))
data <- data[sameSample, ]
cli <- cli[sameSample, ]
out <- cbind(cli, data)
out <- cbind(id = row.names(out), out)
write.table(out, file = "expTime.txt", sep = "\t", row.names = F, quote = F)
