==> 19.R <==
library(survival)
library(survminer)
library(timeROC)

riskFile <- "risk.all.txt"
cliFile <- "clinical.txt"
setwd("C:\\Users\\Zhang TONGTONG\\Desktop\\3\\19-1")

risk <- read.table(riskFile, header = T, sep = "\t", check.names = F, row.names = 1)
risk <- risk[, c("futime", "fustat", "riskScore")]

cli <- read.table(cliFile, header = T, sep = "\t", check.names = F, row.names = 1)

samSample <- intersect(row.names(risk), row.names(cli))
risk1 <- risk[samSample, , drop = F]
cli <- cli[samSample, , drop = F]
rt <- cbind(risk1, cli)

bioCol <- rainbow(ncol(rt) - 1, s = 0.9, v = 0.9)

ROC_rt <- timeROC(
    T = risk$futime, delta = risk$fustat,
    marker = risk$riskScore, cause = 1,
    weighting = "aalen",
    times = c(1, 3, 5), ROC = TRUE
)
pdf(file = "ROC.pdf", width = 5, height = 5)
plot(ROC_rt, time = 1, col = bioCol[1], title = FALSE, lwd = 2)
plot(ROC_rt, time = 3, col = bioCol[2], add = TRUE, title = FALSE, lwd = 2)
plot(ROC_rt, time = 5, col = bioCol[3], add = TRUE, title = FALSE, lwd = 2)
legend("bottomright",
    c(
        paste0("AUC at 1 years: ", sprintf("%.03f", ROC_rt$AUC[1])),
        paste0("AUC at 3 years: ", sprintf("%.03f", ROC_rt$AUC[2])),
        paste0("AUC at 5 years: ", sprintf("%.03f", ROC_rt$AUC[3]))
    ),
    col = bioCol[1:3], lwd = 2, bty = "n"
)
dev.off()

predictTime <- 1
aucText <- c()
pdf(file = "cliROC.pdf", width = 5, height = 5)
i <- 3
ROC_rt <- timeROC(
    T = risk$futime,
    delta = risk$fustat,
    marker = risk$riskScore, cause = 1,
    weighting = "aalen",
    times = c(predictTime), ROC = TRUE
)
plot(ROC_rt, time = predictTime, col = bioCol[i - 2], title = FALSE, lwd = 2)
aucText <- c(paste0("Risk", ", AUC=", sprintf("%.3f", ROC_rt$AUC[2])))
abline(0, 1)
for (i in 4:ncol(rt)) {
    ROC_rt <- timeROC(
        T = rt$futime,
        delta = rt$fustat,
        marker = rt[, i], cause = 1,
        weighting = "aalen",
        times = c(predictTime), ROC = TRUE
    )
    plot(ROC_rt, time = predictTime, col = bioCol[i - 2], title = FALSE, lwd = 2, add = TRUE)
    aucText <- c(aucText, paste0(colnames(rt)[i], ", AUC=", sprintf("%.3f", ROC_rt$AUC[2])))
}
legend("bottomright", aucText, lwd = 2, bty = "n", col = bioCol[1:(ncol(rt) - 1)])
dev.off()
