==> 28.R <==
library(limma)
library(estimate)

inputFile <- "symbol.txt"
setwd("C:\\Users\\Zhang TONGTONG\\Desktop\\3\\28")

rt <- read.table(inputFile, header = T, sep = "\t", check.names = F)
rt <- as.matrix(rt)
rownames(rt) <- rt[, 1]
exp <- rt[, 2:ncol(rt)]
dimnames <- list(rownames(exp), colnames(exp))
data <- matrix(as.numeric(as.matrix(exp)), nrow = nrow(exp), dimnames = dimnames)
data <- avereps(data)

group <- sapply(strsplit(colnames(data), "\\-"), "[", 4)
group <- sapply(strsplit(group, ""), "[", 1)
group <- gsub("2", "1", group)
data <- data[, group == 0]
out <- data[rowMeans(data) > 0, ]

out <- rbind(ID = colnames(out), out)
write.table(out, file = "uniq.symbol.txt", sep = "\t", quote = F, col.names = F)

filterCommonGenes(
    input.f = "uniq.symbol.txt",
    output.f = "commonGenes.gct",
    id = "GeneSymbol"
)

estimateScore(
    input.ds = "commonGenes.gct",
    output.ds = "estimateScore.gct",
    platform = "illumina"
)

scores <- read.table("estimateScore.gct", skip = 2, header = T)
rownames(scores) <- scores[, 1]
scores <- t(scores[, 3:ncol(scores)])
rownames(scores) <- gsub("\\.", "\\-", rownames(scores))
out <- rbind(ID = colnames(scores), scores)
write.table(out, file = "TMEscores.txt", sep = "\t", quote = F, col.names = F)
