==> 30.R <==
library(limma)
library(reshape2)
library(ggplot2)
library(ggpubr)

expFile <- "symbol.txt"
riskFile <- "risk.all.txt"
geneFile <- "gene.txt"
setwd("C:\\Users\\Zhang TONGTONG\\Desktop\\3\\30")

rt <- read.table(expFile, header = T, sep = "\t", check.names = F)
rt <- as.matrix(rt)
rownames(rt) <- rt[, 1]
exp <- rt[, 2:ncol(rt)]
dimnames <- list(rownames(exp), colnames(exp))
data <- matrix(as.numeric(as.matrix(exp)), nrow = nrow(exp), dimnames = dimnames)
data <- avereps(data)

gene <- read.table(geneFile, header = F, sep = "\t", check.names = F)
sameGene <- intersect(row.names(data), as.vector(gene[, 1]))
data <- t(data[sameGene, ])
data <- log2(data + 1)

group <- sapply(strsplit(row.names(data), "\\-"), "[", 4)
group <- sapply(strsplit(group, ""), "[", 1)
group <- gsub("2", "1", group)
data <- data[group == 0, ]
row.names(data) <- gsub("(.*?)\\-(.*?)\\-(.*?)\\-(.*?)\\-.*", "\\1\\-\\2\\-\\3", row.names(data))
data <- avereps(data)

risk <- read.table(riskFile, sep = "\t", header = T, check.names = F, row.names = 1)
sameSample <- intersect(row.names(data), row.names(risk))
rt1 <- cbind(data[sameSample, ], risk[sameSample, ])
rt1 <- rt1[, c(sameGene, "risk")]

sigGene <- c()
for (i in colnames(rt1)[1:(ncol(rt1) - 1)]) {
    if (sd(rt1[, i]) < 0.001) {
        next
    }
    wilcoxTest <- wilcox.test(rt1[, i] ~ rt1[, "risk"])
    pvalue <- wilcoxTest$p.value
    if (wilcoxTest$p.value < 0.05) {
        sigGene <- c(sigGene, i)
    }
}
sigGene <- c(sigGene, "risk")
rt1 <- rt1[, sigGene]

rt1 <- melt(rt1, id.vars = c("risk"))
colnames(rt1) <- c("risk", "Gene", "Expression")

group <- levels(factor(rt1$risk))
rt1$risk <- factor(rt1$risk, levels = c("low", "high"))
comp <- combn(group, 2)
my_comparisons <- list()
for (j in 1:ncol(comp)) {
    my_comparisons[[j]] <- comp[, j]
}

boxplot <- ggboxplot(rt1,
    x = "Gene", y = "Expression", fill = "risk",
    xlab = "",
    ylab = "Gene expression",
    legend.title = "Risk",
    width = 0.8,
    palette = c("#0066FF", "#FF0000")
) +
    rotate_x_text(50) +
    stat_compare_means(aes(group = risk),
        method = "wilcox.test",
        symnum.args = list(cutpoints = c(0, 0.001, 0.01, 0.05, 1), symbols = c("***", "**", "*", "ns")), label = "p.signif"
    )

pdf(file = "checkpoint.diff.pdf", width = 8, height = 5)
print(boxplot)
dev.off()
